+++ packages/core/tests/Axis.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react'
-import { Chart, Axis, AxisTicks, View } from '../src'
+import { Chart, Axis, View } from '../src'
 import { chartProps, viewProps } from './props'
 
 describe('Axis', () => {
@@ -109,78 +109,5 @@ describe('Axis', () => {
         // the chart inner width is 400 - 40 -40 = 320
         // label at the end of the axis means a transform with translate(320
         expect(label?.getAttribute('style')).toContain('translateX(320')
-    })
-})
-
-describe('AxisTicks', () => {
-    it('formats tick labels', () => {
-        const percentFormat = (v: unknown) => v + '%'
-        render(
-            <Chart {...chartProps}>
-                <View {...viewProps}>
-                    <Axis variant="top">
-                        <AxisTicks variant="top" ticks={6} labelFormat={percentFormat} />
-                    </Axis>
-                </View>
-            </Chart>
-        )
-        const result = screen.getAllByRole('tick-label')
-        expect(result[0].textContent).toContain('%')
-    })
-
-    it('can omit ticks altogether', () => {
-        render(
-            <Chart {...chartProps}>
-                <View {...viewProps}>
-                    <Axis variant="top">
-                        <AxisTicks variant="top" ticks={[]} />
-                    </Axis>
-                </View>
-            </Chart>
-        )
-        const result = screen.queryAllByRole('axis-ticks')
-        expect(result).toHaveLength(0)
-    })
-
-    it('can omit tick labels', () => {
-        render(
-            <Chart {...chartProps}>
-                <View {...viewProps}>
-                    <Axis variant="top">
-                        <AxisTicks variant="top" ticks={6} labelFormat={null} />
-                    </Axis>
-                </View>
-            </Chart>
-        )
-        const result = screen.queryAllByRole('tick-label')
-        expect(result).toHaveLength(0)
-    })
-
-    it('can request specific ticks', () => {
-        render(
-            <Chart {...chartProps}>
-                <View {...viewProps}>
-                    <Axis variant="top">
-                        <AxisTicks variant="top" ticks={[0, 20, 30]} />
-                    </Axis>
-                </View>
-            </Chart>
-        )
-        const result = screen.queryAllByRole('tick-label')
-        expect(result).toHaveLength(3)
-    })
-
-    it('can rotate tick labels', () => {
-        render(
-            <Chart {...chartProps}>
-                <View {...viewProps}>
-                    <Axis variant="top">
-                        <AxisTicks variant="top" ticks={[0, 50, 100]} labelRotate={45} />
-                    </Axis>
-                </View>
-            </Chart>
-        )
-        const result = screen.queryAllByRole('tick-label')
-        expect(result[1].getAttribute('transform')).toContain('rotate(')
     })
 })